% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{annoSegment}
\alias{annoSegment}
\title{Annotation modifier for plots}
\usage{
annoSegment(
  object = NULL,
  relSideDist = 0.1,
  aesGroup = FALSE,
  aesGroName = NULL,
  annoPos = "top",
  xPosition = NULL,
  yPosition = NULL,
  pCol = NULL,
  segWidth = 1,
  lty = NULL,
  lwd = 10,
  alpha = NULL,
  lineend = "square",
  annoManual = FALSE,
  mArrow = NULL,
  addBranch = FALSE,
  bArrow = NULL,
  branDirection = 1,
  branRelSegLen = 0.3,
  addText = FALSE,
  textCol = NULL,
  textSize = NULL,
  fontfamily = NULL,
  fontface = NULL,
  textLabel = NULL,
  textRot = 0,
  textHVjust = 0.2,
  hjust = NULL,
  vjust = NULL,
  myFacetGrou = NULL,
  aes_x = NULL,
  aes_y = NULL
)
}
\arguments{
\item{object}{ggplot list. Default(NULL).}

\item{relSideDist}{The relative distance ratio to the y axis range. Default(0.1).}

\item{aesGroup}{Whether use your group column to add rect annotation. Default("FALSE").}

\item{aesGroName}{The mapping column name. Default(NULL).}

\item{annoPos}{The position for the annotation to be added. Default("top").}

\item{xPosition}{The x axis coordinate for the segment. Default(NULL).}

\item{yPosition}{The y axis coordinate for the segment. Default(NULL).}

\item{pCol}{The segment colors. Default(NULL).}

\item{segWidth}{The relative segment width. Default(1).}

\item{lty}{The segment line type. Default(NULL).}

\item{lwd}{The segment line width. Default(NULL).}

\item{alpha}{The segment color alpha. Default(NULL).}

\item{lineend}{The segment line end. Default("square").}

\item{annoManual}{Whether annotate by yourself by supplying with x and y coordinates. Default(FALSE).}

\item{mArrow}{Whether add segment arrow. Default(FALSE).}

\item{addBranch}{Whether add segment branch. Default(FALSE).}

\item{bArrow}{Whether add branch arrow. Default(FALSE).}

\item{branDirection}{The branch direction. Default(1).}

\item{branRelSegLen}{The branch relative length to the segment. Default(0.3).}

\item{addText}{Whether add text label on segment. Default(FALSE).}

\item{textCol}{The text colors. Default(NULL).}

\item{textSize}{The text size. Default(NULL).}

\item{fontfamily}{The text fontfamily. Default(NULL).}

\item{fontface}{The text fontface. Default(NULL).}

\item{textLabel}{The text textLabel. Default(NULL).}

\item{textRot}{The text angle. Default(NULL).}

\item{textHVjust}{The text distance from the segment. Default(0.2).}

\item{hjust}{The text hjust. Default(NULL).}

\item{vjust}{The text vjust. Default(NULL).}

\item{myFacetGrou}{Your facet group name to be added with annotation when object is a faceted object. Default(NULL).}

\item{aes_x}{= NULL You should supply the plot X mapping name when annotate a facetd plot. Default(NULL).}

\item{aes_y}{= NULL You should supply the plot Y mapping name when annotate a facetd plot. Default(NULL).}
}
\description{
Used for segment the plot for further annotations
}
\author{
Jun Zhang
}
