% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilisk.R
\docType{data}
\name{cellbender_env}
\alias{cellbender_env}
\title{Cellbender Python Environment}
\format{
An object of class \code{BasiliskEnvironment} of length 1.
}
\usage{
cellbender_env
}
\value{
A \code{\link[basilisk:BasiliskEnvironment-class]{basilisk::BasiliskEnvironment()}} object preconfigured with Cellbender.
}
\description{
A \code{BasiliskEnvironment} containing Cellbender and Python 3.7, used in the DOtools package for running Cellbender-related functionality.
}
\keyword{datasets}
