% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{DO.BarcodeRanks}
\alias{DO.BarcodeRanks}
\title{Run DropletUtils::barcodeRanks to estimate cell bounds}
\usage{
DO.BarcodeRanks(SCE_obj)
}
\arguments{
\item{matrix}{A sparse count matrix of class "dgCMatrix".}
}
\value{
A named numeric vector: \code{c(knee = ..., inflection = ...)}
}
\description{
Given a raw count matrix (e.g. from a CellRanger HDF5 file), estimate the number of expected cells and droplets
using the knee and inflection points from barcodeRanks.
}
