% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{DO.Dotplot}
\alias{DO.Dotplot}
\title{DO Dot plot}
\usage{
DO.Dotplot(
  Seu_object,
  Feature,
  group.by.x = NULL,
  group.by.y = NULL,
  group.by.y2 = NULL,
  across.group.by.x = F,
  dot.size = c(1, 6),
  plot.margin = c(1, 1, 1, 1),
  midpoint = 0.5,
  colZ = F,
  returnValue = F,
  log1p_nUMI = T,
  hide_zero = T,
  annotation_x = F,
  annotation_x_position = 0.25,
  annotation_x_rev = F,
  point_stroke = 0.2,
  limits_colorscale = NULL,
  coord_flip = F,
  ...
)
}
\arguments{
\item{Seu_object}{The seurat Seu_object}

\item{Feature}{Genes or DF of interest, Data frame should have columns with gene and annotation information, e.g. output of FindAllMarkers}

\item{group.by.x}{group name to plot on x-axis}

\item{group.by.y}{group name to look for in meta data}

\item{group.by.y2}{second group name to look for in meta data}

\item{across.group.by.x}{calculate a pseudobulk expression approach for the x-axis categories}

\item{dot.size}{Vector of dot size}

\item{plot.margin}{= plot margins}

\item{midpoint}{midpoint in color gradient}

\item{colZ}{IF True calculates the Z-score of the average expression per column}

\item{hide_zero}{Removes dots for genes with 0 expression}

\item{annotation_x}{Adds annotation on top of x axis instead on y axis}

\item{point_stroke}{Defines the thickness of the black stroke on the dots}

\item{limits_colorscale}{Set manually colorscale limits}

\item{...}{Further arguments passed to annoSegment function if annotation_x == T}
}
\value{
a ggplot
}
\description{
This function generates a dot plot for multiple genes, comparing expression levels across one or two specified groups.
It supports both individual and pseudobulk expression calculations.
Highly variable customization options allow control over dot size, color scaling, annotations, and axis orientation.
The function integrates seamlessly with Seurat objects for single-cell RNA-seq analysis.
}
\examples{
\dontrun{


DO.Dotplot(
  Seu_object = Seurat,
  Feature = c("CDH5","TTN","MALAT1"),
  group.by.x="condition"
)
}


}
\author{
Mariano Ruz Jurado
}
