% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{DO.Mean.SEM.Graphs.wilcox}
\alias{DO.Mean.SEM.Graphs.wilcox}
\title{SEM Graph with wilcox test on single cell level}
\usage{
DO.Mean.SEM.Graphs.wilcox(
  Seu_object,
  Feature,
  ListTest = NULL,
  returnValues = FALSE,
  ctrl.condition = NULL,
  group.by = "condition",
  wilcox_test = TRUE,
  bar_colours = NULL,
  stat_pos_mod = 1.15,
  x_label_rotation = 45,
  plotPvalue = FALSE,
  y_limits = NULL,
  log1p_nUMI = T
)
}
\arguments{
\item{Seu_object}{combined Seurat object}

\item{Feature}{name of the feature/gene}

\item{ListTest}{List for which conditions wilcoxon test will be performed, if NULL always CTRL group against everything}

\item{returnValues}{return data frames needed for the plot, containing df.melt, df.melt.sum, df.melt.orig and wilcoxstats}

\item{ctrl.condition}{set your ctrl condition, relevant if running with empty comparison List}

\item{group.by}{select the seurat object slot where your conditions can be found, default conditon}

\item{bar_colours}{colour vector}

\item{plotPvalue}{plot the non adjusted p-value without correcting for multiple tests}

\item{SeuV5}{Seuratv5 object? (TRUE or FALSE)}
}
\description{
Perform SEM-based graphs with Wilcox test on single-cell level for Seurat objects.
Calculates mean expression values and SEM for the selected feature, and visualizes them.
Performs pairwise Wilcox tests comparing conditions, with optional custom control condition and clustering.
Optionally returns a summary data frame, statistical test results, and the generated plot.
}
\author{
Mariano Ruz Jurado
}
