% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{DO.PyEnv}
\alias{DO.PyEnv}
\title{DO.PyEnv}
\usage{
DO.PyEnv(conda_path = NULL)
}
\arguments{
\item{conda_path}{character string specifying the path to an existing or new conda environment.}
}
\description{
Sets up or connects to a conda Python environment for use with DOtools.
If no environment path is provided, it will create one at \verb{~/.venv/DOtools} and install required Python packages:
\code{scvi-tools}, \code{celltypist}, and \code{scanpro}.
}
\examples{
\dontrun{
# Automatically create DOtools environment at ~/.venv/DOtools if it doesn't exist
DO.PyEnv()

# Use an existing conda environment at a custom location
DO.PyEnv(conda_path = "~/miniconda3/envs/my_dotools_env")
}


}
