% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO.UMAP.R
\name{DO.UMAP}
\alias{DO.UMAP}
\title{DO.UMAP}
\usage{
DO.UMAP(
  sce_object,
  FeaturePlot = FALSE,
  DensityPlot = FALSE,
  features = NULL,
  reduction = NULL,
  group.by = "seurat_clusters",
  dims = c(1, 2),
  layer = NULL,
  umap_colors = NULL,
  text_size = 14,
  label = TRUE,
  order = TRUE,
  plot.title = TRUE,
  legend.position = "none",
  method = c("ks", "wkde"),
  ...
)
}
\arguments{
\item{sce_object}{The seurat or SCE object}

\item{FeaturePlot}{Is it going to be a FeaturePlot?}

\item{DensityPlot}{Is it going to be a DensityPlot?}

\item{features}{features for Featureplot}

\item{reduction}{reduction to use}

\item{group.by}{grouping of plot in DImplot and defines in featureplot the
labels}

\item{dims}{Dimensions to plot, must be a two-length numeric vector
specifying x- and y-dimensions description}

\item{layer}{Layer to use for DensityPlot, default data}

\item{umap_colors}{what colors to use for UMAP, specify as vector}

\item{text_size}{Size of text}

\item{label}{label the clusters on the plot by group.by column}

\item{order}{Boolean determining whether to plot cells in order of
expression.}

\item{plot.title}{title for UMAP}

\item{legend.position}{specify legend position}

\item{method}{Kernel density estimation method, can be "ks" or "wkde"}

\item{...}{Further arguments passed to DimPlot, FeaturePlot or DensityPlot
functions}
}
\value{
Plot with Refined colors and axes
}
\description{
Creates a polished UMAP plot using Seurat's DimPlot or
FeaturePlot functions. In addition a density plot can be made in a nebulosa
R package way. It allows customization of colors, labels, and other
plot elements for better visualisation. The function handles both
cluster-based visualisations and gene-based visualisations in a UMAP plot.
Ideal for refining UMAP outputs with added flexibility and enhanced
presentation.
}
\examples{
sce_data <-
    readRDS(system.file("extdata", "sce_data.rds", package = "DOtools"))

DO.UMAP(
    sce_object = sce_data,
    group.by = "seurat_clusters"
)

DO.UMAP(
    sce_object = sce_data,
    FeaturePlot = TRUE,
    features = c("BAG2", "CD74")
)

DO.UMAP(
    sce_object = sce_data,
    DensityPlot = TRUE,
    features = c("CD74")
)

}
\author{
Mariano Ruz Jurado
}
