% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{DO.Vln.Plot.wilcox}
\alias{DO.Vln.Plot.wilcox}
\title{Violin Graph with wilcox test on single cell level}
\usage{
DO.Vln.Plot.wilcox(
  Seu_object,
  SeuV5 = T,
  Feature,
  ListTest = NULL,
  returnValues = FALSE,
  ctrl.condition = NULL,
  group.by = "condition",
  group.by.2 = NULL,
  geom_jitter_args = c(0.2, 0.25, 0.25),
  geom_jitter_args_group_by2 = c(0.1, 0.1, 1),
  vector_colors = c("#1f77b4", "#ea7e1eff", "royalblue4", "tomato2", "darkgoldenrod",
    "palegreen4", "maroon", "thistle3"),
  wilcox_test = T,
  stat_pos_mod = 1.15,
  hjust.wilcox = 0.8,
  vjust.wilcox = 2,
  size.wilcox = 3.33,
  step_mod = 0,
  hjust.wilcox.2 = 0.5,
  vjust.wilcox.2 = 0,
  width_errorbar = 0.4
)
}
\arguments{
\item{Seu_object}{combined Seurat object}

\item{SeuV5}{Seuratv5 object? (TRUE or FALSE)}

\item{Feature}{name of the feature}

\item{ListTest}{List for which conditions wilcox will be performed, if NULL always CTRL group against everything}

\item{returnValues}{return df.melt.sum data frame containing means and SEM for the set group}

\item{ctrl.condition}{set your ctrl condition, relevant if running with empty comparison List}

\item{group.by}{select the seurat Seu_object slot where your conditions can be found, default conditon}

\item{group.by.2}{relevant for multiple group testing, e.g. for each cell type the test between each of them in two conditions provided}

\item{geom_jitter_args}{vector for dots visualisation in vlnplot: size, width, alpha value}

\item{wilcox_test}{Bolean if TRUE a bonferoni wilcoxon test will be carried out between ctrl.condition and the rest}

\item{stat_pos_mod}{value for modifiyng statistics height}

\item{vector_colours}{specify a minimum number of colours as you have entries in your condition, default 2}
}
\description{
Creates a violin plot to compare gene expression across different conditions or groups within a Seurat object.
It incorporates Wilcoxon rank-sum tests to evaluate statistical differences between conditions.
The plot can be customized with options for data transformation, jitter display, and significance annotations.
The function also supports multiple conditions and allows for visualization of statistical results from wilcoxon test.
}
\author{
Mariano Ruz Jurado
}
