% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{.DO.BarcodeRanks}
\alias{.DO.BarcodeRanks}
\title{DO.BarcodeRanks}
\usage{
.DO.BarcodeRanks(SCE_obj)
}
\arguments{
\item{SCE_object}{A Single cell experiment object.}
}
\value{
A named numeric vector: \code{c(xpc_cells = ..., total_cells = ...)}
}
\description{
Given a raw count matrix (e.g. from a CellRanger HDF5 file), estimate the number of expected cells and droplets
using the knee and inflection points from barcodeRanks.
}
\author{
Mariano Ruz Jurado
}
\keyword{internal}
