% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO.UMAP.R
\name{.plot_density_}
\alias{.plot_density_}
\title{Plot density estimates}
\usage{
.plot_density_(
  z,
  dens,
  feature,
  cell_embeddings,
  dim_names,
  shape = 16,
  size = 1,
  text_size = 14,
  density_quantile_threshold = 0.5,
  n_bands = 256,
  legend.position = "right",
  legend_title = "Density",
  pal = "Reds",
  raster = FALSE,
  ...
)
}
\arguments{
\item{z}{Vector with density values for each cells}

\item{dens}{density grid retrieved from KDE or wkde2d}

\item{feature}{Name of the feature being plotted}

\item{cell_embeddings}{Matrix with cell embeddings}

\item{dim_names}{Names of the dimensions from the cell embeddings}

\item{shape}{Geom shape}

\item{size}{Geom size}

\item{text_size}{text_size given from parent function}

\item{legend.position}{legend.position from parent function}

\item{legend_title}{String used as legend title}

\item{pal}{String specifying the color palette to use, can be one of hcl.pals}

\item{raster}{Rasterise plot}

\item{...}{Further scale arguments passed to scale_color_viridis_c}
}
\value{
A ggplot object
}
\description{
Plot density estimates
}
\author{
Jose Alquicira-Hernandez (modified)
}
\keyword{internal}
