% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilisk.R
\docType{data}
\name{dotools_env}
\alias{dotools_env}
\title{DOtools Python Environment}
\format{
An object of class \code{BasiliskEnvironment} of length 1.
}
\usage{
dotools_env
}
\value{
A \code{\link[basilisk:BasiliskEnvironment-class]{basilisk::BasiliskEnvironment()}} object preconfigured with DOtools-related packages.
}
\description{
A \code{BasiliskEnvironment} containing general-purpose Python packages used in DOtools, including celltypist and scanpro, with Python 3.11.
}
\keyword{datasets}
