% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO.EvalIntegration.R
\name{run_kbet}
\alias{run_kbet}
\title{kBET function}
\usage{
run_kbet(
  sce_object,
  embedding,
  batch,
  cells.use = NULL,
  subsample = NULL,
  min_per_batch = NULL,
  ...
)
}
\arguments{
\item{sce_object}{Seurat or SCE object.}

\item{embedding}{Name of the embedding to test}

\item{batch}{Name of the sample column in meta.data}

\item{cells.use}{specify a specific set of cells as vector}

\item{subsample}{set a fraction for stratified subsetting}

\item{min_per_batch}{should always be higher than or equal k0}
}
\value{
list with summary stats of kBET
}
\description{
kBET quantifies batch mixing in single-cell data by testing
whether local neighborhood batch composition deviates from the global
distribution, with lower rejection indicating better integration.
}
\author{
Mariano Ruz Jurado
}
\keyword{internal}
