% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{DO.BarplotClustert}
\alias{DO.BarplotClustert}
\title{SEM Graph with t test on cluster level}
\usage{
DO.BarplotClustert(
  Seu_object,
  Features,
  ListTest = NULL,
  returnValues = FALSE,
  ctrl.condition = NULL,
  group.by = "condition",
  returnPlot = FALSE
)
}
\arguments{
\item{Seu_object}{Combined Seu_object}

\item{Features}{Vector containing featurenames}

\item{ListTest}{List for which conditions t-test will be performed, if NULL always against provided CTRL}

\item{returnValues}{return df.melt.sum data frame containing means and SEM for the set group}

\item{ctrl.condition}{set your ctrl condition, relevant if running with empty comparison List}

\item{group.by}{select the seurat object slot where your conditions can be found, default conditon}
}
\value{
a ggplot or a dataframe
}
\description{
Perform SEM-based graphs with t-test on cluster level for Seurat objects.
Calculates mean expression values and SEM for selected features, and visualizes them.
Performs pairwise t-tests comparing conditions, with optional custom control condition and clustering.
Optionally returns a summary data frame.
}
\examples{
\dontrun{

(optional)
ListTest <- list()
ListTest[[1]] <- c("CTRL", "CONDITION")

DO.BarplotClustert(
  Seu_object = Seurat,
  Features = "CDH5",
  ListTest = ListTest,
  ctrl.condition = "CTRL",
  group.by="condition"
)
}

}
\author{
Mariano Ruz Jurado
}
