% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{DO.FullRecluster}
\alias{DO.FullRecluster}
\title{DO.FullRecluster}
\usage{
DO.FullRecluster(
  SeuratObject,
  res = 0.5,
  algorithm = 4,
  graph.name = "RNA_snn"
)
}
\arguments{
\item{SeuratObject}{The seurat object}

\item{res}{Resolution for the new clusters, default 0.5}

\item{algorithm}{Set one of the available algorithms found in FindSubCLuster function, default = 4: leiden}
}
\value{
Seurat Object with new clustering named eurat_Recluster
}
\description{
Creates a refined clustering for each major cluster found initially by FindClusters
}
\author{
Mariano Ruz Jurado
}
