% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{DO.Subset}
\alias{DO.Subset}
\title{DO.Subset}
\usage{
DO.Subset(
  SeuratObject,
  assay = "RNA",
  ident,
  ident_name = NULL,
  ident_thresh = NULL
)
}
\arguments{
\item{SeuratObject}{The seurat object}

\item{assay}{assay to subset by}

\item{ident}{meta data column to subset for}

\item{ident_name}{name of group of barcodes in ident of subset for}

\item{ident_thresh}{numeric thresholds as character, e.g ">5" or c(">5", "<200"), to subset barcodes in ident for}
}
\value{
subsetted Seurat object
}
\description{
Creates a subset from your previous Seurat object with the ident and name you want to subset
}
\author{
Mariano Ruz Jurado
}
