% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{DO.UMAP}
\alias{DO.UMAP}
\title{DO.UMAP}
\usage{
DO.UMAP(
  SeuratObject,
  FeaturePlot = F,
  features = NULL,
  group.by = "seurat_clusters",
  umap_colors = NULL,
  text_size = 14,
  label = T,
  order = T,
  plot.title = T,
  legend.position = "none",
  ...
)
}
\arguments{
\item{SeuratObject}{The seurat object}

\item{FeaturePlot}{Is it going to be a Dimplot or a FeaturePlot?}

\item{features}{features for Featureplot}

\item{group.by}{grouping of plot in DImplot and defines in featureplot the labels}

\item{...}{Further arguments passed to DimPlot or FeaturePlot function from Seurat}
}
\value{
Plot with Refined colors and axes
}
\description{
Creates a polished UMAP from Dimplot or FeaturePlot function
}
\author{
Mariano Ruz Jurado
}
