% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{DO.UMAP}
\alias{DO.UMAP}
\title{DO.UMAP}
\usage{
DO.UMAP(
  Seu_object,
  FeaturePlot = F,
  features = NULL,
  group.by = "seurat_clusters",
  umap_colors = NULL,
  text_size = 14,
  label = T,
  order = T,
  plot.title = T,
  legend.position = "none",
  ...
)
}
\arguments{
\item{Seu_object}{The seurat object}

\item{FeaturePlot}{Is it going to be a Dimplot or a FeaturePlot?}

\item{features}{features for Featureplot}

\item{group.by}{grouping of plot in DImplot and defines in featureplot the labels}

\item{...}{Further arguments passed to DimPlot or FeaturePlot function from Seurat}
}
\value{
Plot with Refined colors and axes
}
\description{
Creates a polished UMAP plot using Seurat's DimPlot or FeaturePlot functions.
It allows customization of colors, labels, and other plot elements for better visualization.
The function handles both cluster-based visualizations and gene-based visualizations in a UMAP plot.
Ideal for refining UMAP outputs with added flexibility and enhanced presentation.
}
\examples{
\dontrun{


DO.UMAP(
  Seu_object = Seurat,
  group.by="seurat_clusters"
)

DO.UMAP(
  Seu_object = Seurat,
  FeaturePlot=T,
  features=c("CDH5","TTN")
)
}

}
\author{
Mariano Ruz Jurado
}
